--- spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/security/oauth2/client/reactive/ReactiveOAuth2ClientRegistrationRepositoryConfigurationTests.java
@@ -1,65 +1,0 @@
-/*
- * Copyright 2012-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.boot.autoconfigure.security.oauth2.client.reactive;
-
-import org.junit.Test;
-
-import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.registration.ReactiveClientRegistrationRepository;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * Tests for {@link ReactiveOAuth2ClientRegistrationRepositoryConfiguration}.
- *
- * @author Madhura Bhave
- */
-public class ReactiveOAuth2ClientRegistrationRepositoryConfigurationTests {
-
-	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner();
-
-	private static final String REGISTRATION_PREFIX = "spring.security.oauth2.client.registration.login";
-
-	@Test
-	public void clientRegistrationRepositoryBeanShouldNotBeCreatedWhenPropertiesAbsent() {
-		this.contextRunner
-				.withUserConfiguration(
-						ReactiveOAuth2ClientRegistrationRepositoryConfiguration.class)
-				.run((context) -> assertThat(context)
-						.doesNotHaveBean(ClientRegistrationRepository.class));
-	}
-
-	@Test
-	public void clientRegistrationRepositoryBeanShouldBeCreatedWhenPropertiesPresent() {
-		this.contextRunner
-				.withUserConfiguration(
-						ReactiveOAuth2ClientRegistrationRepositoryConfiguration.class)
-				.withPropertyValues(REGISTRATION_PREFIX + ".foo.client-id=abcd",
-						REGISTRATION_PREFIX + ".foo.client-secret=secret",
-						REGISTRATION_PREFIX + ".foo.provider=github")
-				.run((context) -> {
-					ReactiveClientRegistrationRepository repository = context
-							.getBean(ReactiveClientRegistrationRepository.class);
-					ClientRegistration registration = repository
-							.findByRegistrationId("foo").block();
-					assertThat(registration).isNotNull();
-					assertThat(registration.getClientSecret()).isEqualTo("secret");
-				});
-	}
-
-}
